updated stories
