+++ packages/react-toast/src/lib/index.ts
@@ -1,6 +1,18 @@
+
 import {
   Bounce, Slide, Zoom,
 } from "react-toastify";
+
+export function mergeConfig(defaultConfig, config) {
+  return {
+    ...defaultConfig,
+    ...config,
+    theme: {
+      ...defaultConfig?.theme,
+      ...config?.theme
+    }
+  }
+}
 
 export function getTransition(type?: string) {
   switch (type) {
