--- packages/storybook/stories/react-inbox/Indicator.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Indicator } from "@trycourier/react-inbox";
-
-export default {
-  title: "Inbox/Indicator",
-  component: Indicator,
-  argTypes: {},
-  args: {},
-};
-
-export function Default() {
-  return <Indicator />;
-}
