--- packages/storybook/stories/react-inbox/styled.js
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const Button = styled.button`
-  outline: none;
-  border: none;
-  border-radius: 5px;
-  height: 50px;
-  width: 200px;
-  background-color: black;
-  color: white;
-  font-size: 20px;
-  font-family: Verdana;
-  cursor: pointer;
-`;
