+++ packages/storybook/stories/style.css
@@ -1,0 +1,3 @@
+button {
+  height: 45px;
+}
